move breakpoitns config into style tokens file and reassign import statmements
