+++ app-web/src/components/Home/Masthead.js
@@ -21,7 +21,8 @@ import { navigate } from 'gatsby';
 import styled from '@emotion/styled';
 import { css } from '@emotion/core';
 import { ChevronLink } from '../UI/Link';
-import { SEARCH, EMOTION_BOOTSTRAP_BREAKPOINTS } from '../../constants/ui';
+import { SEARCH } from '../../constants/ui';
+import { EMOTION_BOOTSTRAP_BREAKPOINTS } from '../../constants/styleTokens';
 import Search from '../Search';
 import SearchPills from '../Search/SearchPills';
 // localizations
