+++ app-web/src/components/Page/FilterMenu.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import Filters from '../Filters/Filters';
-import { EMOTION_BOOTSTRAP_BREAKPOINTS } from '../../constants/ui';
+import { EMOTION_BOOTSTRAP_BREAKPOINTS } from '../../constants/styleTokens';
 const FilterContainer = styled.div`
   flex: 0 0 200px;
   padding: 5px;
