+++ app-web/src/constants/ui.js
@@ -50,31 +50,6 @@ export const AVATAR_NAMESPACES = {
   github: 'github',
 };
 
-// bootstrap breakpoints used for emotion custom styled components
-export const BOOTSTRAP_BREAKPOINTS = [
-  { alias: 'sm', width: 576 },
-  { alias: 'md', width: 768 },
-  { alias: 'lg', width: 992 },
-  { alias: 'xl', width: 1200 },
-];
-
-/**
- * returns an object aliased by bootstrap breakpoint values
- * {
- *   sm: @media (min-width: ...)
- *   md: ...
- * }
- * Usage:
- * https://emotion.sh/docs/media-queries#reusable-media-queries
- */
-export const EMOTION_BOOTSTRAP_BREAKPOINTS = BOOTSTRAP_BREAKPOINTS.reduce(
-  (breakpoints, currentBp) => {
-    breakpoints[currentBp.alias] = `@media (min-width: ${currentBp.width}px)`;
-    return breakpoints;
-  },
-  {},
-);
-
 export const EVENTS = {
   MAX_PAST_EVENTS: 9,
 };
